+++ docs/News.rst
@@ -14,6 +14,9 @@ Drivers
 * Extended default list of MySQL drivers to ``mysqldb``, ``mysqlclient``,
   ``mysql-connector``, ``mysql-connector-python``, ``pymysql``.
 
+* Extended default list of PostgreSQL drivers to ``psycopg``, ``psycopg2``,
+  ``pygresql``.
+
 SQLObject 3.12.0.post2
 ======================