+++ docs/News.rst
@@ -19,7 +19,7 @@ Drivers
 
 * Fixed outstanding problems with ``psycopg``. It's now the first class driver.
 
-* Dropped support for CyMySQL;
+* Dropped support for ``CyMySQL``;
   its author refused to fix unicode-related problems.
 
 Tests